// frontend/src/api/generated/models/outbound-slip-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 受注ステータス - ORDERED: 受注済み - PARTIAL_ALLOCATED: 一部引当済み - ALLOCATED: 引当完了 - PICKING_COMPLETED: ピッキング完了 - INSPECTING: 検品中 - SHIPPED: 出荷完了 - CANCELLED: キャンセル 
 * @export
 * @enum {string}
 */

export const OutboundSlipStatus = {
    Ordered: 'ORDERED',
    PartialAllocated: 'PARTIAL_ALLOCATED',
    Allocated: 'ALLOCATED',
    PickingCompleted: 'PICKING_COMPLETED',
    Inspecting: 'INSPECTING',
    Shipped: 'SHIPPED',
    Cancelled: 'CANCELLED'
} as const;

export type OutboundSlipStatus = typeof OutboundSlipStatus[keyof typeof OutboundSlipStatus];